Create types and rafactor the App page by creating a separe MainPanel component
