+++ src/lib/enums/index.enum.ts
@@ -1,0 +1,11 @@
+export enum JobTypes {
+  FULL_TIME = "Full-time",
+  PART_TIME = "Part-time",
+  CONTRACT = "Contract",
+}
+
+export enum Workplaces {
+  REMOTE = "Remote",
+  ONSITE = "Onsite",
+  HYBRID = "Hybrid",
+}
